Former-commit-id: 39f944598963368c1348ed63e1e01f5098ae59de
Former-commit-id: c30859fe901daa13c214118232d35bf6c33319ce

+++ src/fswAlgorithms/imageProcessing/houghCircles/houghCircles.h
@@ -70,7 +70,7 @@ public:
     int32_t houghMinRadius;              //!< [-] Min radius of a detected circle
     int32_t houghMaxRadius;              //!< [-] Max radius of a detected circle
     int32_t dpValue;                     //!< [-] Subscaling of image for circle searching, 1 searches full image
-    int32_t noiseSF;                      //!< [-] Scale Factor for noise control
+    double noiseSF;                      //!< [-] Scale Factor for noise control
     int32_t expectedCircles;             //!< [-] Number of expected circles to be found
     int32_t saveImages;                  //!< [-] 1 to save images to file for debugging
 private:
